refactor: split table component
